utils/requirements.txt
scikit-learn~=0.22.1
matplotlib~=3.0.3
keras~=2.2.4
h5py~=2.9.0
pandas~=0.24.2
numpy~=1.16.2
scipy~=1.4.1
tensorflow-gpu~=1.13.1
# keras_contrib # installed using pip3 install git+https://www.github.com/keras-team/keras-contrib.git
pydot~=1.4.1
pydot-ng~=2.0.0
graphviz~=0.13.2
kapre==0.1.4
hyperopt
filelock
jupyter
opencv-python
imbalanced-learn
seaborn
